add test for repository and util classes. fix userAnswerParser

+++ src/main/java/ru/alexsumin/weightstatbot/repository/MeasurementRepository.java
@@ -7,7 +7,6 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
 import ru.alexsumin.weightstatbot.model.Measurement;
 
-import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.Optional;
 
@@ -29,7 +28,7 @@ public interface MeasurementRepository extends CrudRepository<Measurement, Long>
     @Query(value = "SELECT  m.amount  FROM measurement as m" +
             " JOIN account as a ON  m.ACCOUNT_ACCOUNT_ID = a.ACCOUNT_ID" +
             "             WHERE a.chat_id=:id AND m.DATE_OF_MEASUREMENT > DATE_ADD(current_date, INTERVAL -:period DAY) LIMIT 1; ", nativeQuery = true)
-    Optional<BigDecimal> getFirstMeasurementInPeriod(@Param("id") Long chatId, @Param("period") Integer period);
+    Optional<BigInteger> getFirstMeasurementInPeriod(@Param("id") Long chatId, @Param("period") Integer period);
 
 
     @Transactional
